BasicLU.jl

BasicLU.BLU — Type

mutable struct BLU

Hold BASICLU object. You can access @xstore to set parameters and get output values from BASICLU routines.

BasicLU.get_factors — Method

get_factors(this)

Export LU factors after fresh factorization.

Example

m = 100
B = spdiagm(-1 => ones(m-1), 0 => 4*ones(m), 1 => ones(m-1))
this = BasicLU.initialize(m)
BasicLU.factorize(this, B)
(L,U,p,q) = BasicLU.get_factors(this)
norm(L*U-B[p,q], Inf)
1.1102230246251565e-16

BasicLU.initialize — Method

initialize(m)

Return a new BASICLU object for matrices of dimension m.

BasicLU.maxvolume_basis — Method

basis, obj = maxvolume_basis(A::spmatrix)

Returns a set of column indices for the matrix AI = [A sparse(I, m, m)] such
that AI[:,basis] is nonsingular and the number of slack columns in the basis
is minimum (this is the row rank deficiency of A).

BasicLU.solve — Method

solve(this, rhs, trans)

Solve linear system with factorized matrix.

@rhs must be a dense or sparse vector. The solution is returned sparse when @rhs is sparse and dense otherwise.

@trans must be 't' or 'T' for transposed solve, and any other character for forward solve.

BasicLU.solve4update — Function

solve4update(this, rhs, getsolution=false)

Solve linear system in preparation to update the factorization.

@rhs must be a sparse vector or a column index. When @rhs is a vector, then it is the column to be inserted into the factorized matrix. When @rhs is an index, than it is the column of the factorized matrix to be replaced.

@getsolution indicates if the solution is to be returned, or only the update is to be prepared.

BasicLU.update — Method

update(this, xtbl)

Update the factorization after a column modification. The column position and the new column must have been set in previous calls to @solve4update.

@xtbl is the pivot element in the simplex tableau. Used only to compute the pivot error.

Return estimated error of the new pivot element.

LinearAlgebra.factorize — Method

factorize(this, B)

Load sparse matrix into BASICLU object and factorize it. @B must be square and have the same dimension for which @this was initialized.

Return status code.